Refactor (un)nesting functions into new utils module
